+++ examples/regression/plot_prefit_nn.py
@@ -3,7 +3,7 @@
 Example use of the prefit parameter with neural networks
 ========================================================
 
-:class:`mapie.estimators.MapieRegressor` is used to calibrate
+:class:`mapie.regression.MapieRegressor` is used to calibrate
 uncertainties for large models for which the cost of cross-validation
 is too high. Typically, neural networks rely on a single validation set.
 
@@ -17,8 +17,8 @@ from sklearn.model_selection import train_test_split
 from sklearn.neural_network import MLPRegressor
 from matplotlib import pyplot as plt
 
-from mapie.estimators import MapieRegressor
-from mapie.metrics import coverage_score
+from mapie.regression import MapieRegressor
+from mapie.metrics import regression_coverage_score
 
 
 def f(x: np.ndarray) -> np.ndarray:
@@ -52,7 +52,7 @@ mapie.fit(X_val.reshape(-1, 1), y_val)
 alpha = 0.1
 y_pred, y_pis = mapie.predict(X_test.reshape(-1, 1), alpha=alpha)
 y_pred_low, y_pred_up = y_pis[:, 0, 0], y_pis[:, 1, 0]
-coverage = coverage_score(y_test, y_pred_low, y_pred_up)
+coverage = regression_coverage_score(y_test, y_pred_low, y_pred_up)
 
 # Plot obtained prediction intervals on testing set
 theoretical_semi_width = scipy.stats.norm.ppf(1 - alpha)*sigma