+++ mapie/__init__.py
@@ -1,6 +1,7 @@
-from . import estimators
+from . import regression
+from . import classification
 from . import metrics
 
 from ._version import __version__
 
-__all__ = ["estimators", "metrics", "__version__"]
+__all__ = ["regression", "classification", "metrics", "__version__"]